requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.13.3